requirements.txt
Django==5.2.8
django-crispy-forms==2.3
Django==5.2.8
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
Pillow==10.4.0
djangorestframework==3.15.2
django-filter==24.3
markdown==3.7
django-markdownx==4.0.8
psycopg2-binary
python-decouple==3.8
whitenoise==6.7.0
xhtml2pdf==0.2.16
drf-spectacular==0.27.0